Remove special handling for `impl Trait for ..` syntax errors.
The ancient (pre-1.0) RFC 19 suggested using `impl Trait for ..` syntax
for default traits. That was later changed to `auto trait Trait {}`
syntax. The parser has special treatment for the `..` syntax, suggesting
the `auto` syntax.

Given that default traits have not be stabilized and the `..` syntax is
so old, the special case seems unnecessary, and it gets in the way of
adding `ErrorGuaranteed` to `TyKind::Err`. This commit removes it and
the tests.