Deduplicate oject safety errors on `impl dyn Trait { .. }`
`impl dyn Trait {}` has an implicit `dyn Trait: 'static` bound. We add it in
`inferred_outlives_of` for more context in errors. With this we point at
`impl dyn Trait` and not to it's associated functions, deduplicating errors
to a single one per `impl dyn Trait` instead of one for it plus one for each
associated function.