Return `Option<ErrorGuaranteed>` from `lint_level_impl`.
`lint_level_impl` emits a lint that may be an error or a non-error,
depending on the setting. By introducing `impl EmissionGuarantee for
Option<ErrorGuaranteed>`, we can express this in the type system via its
return value, and the return value of many other linting functions built
on top of `lint_level_impl`.

In `increment_const_eval_counter`, this lets us avoid a
`span_delayed_bug` for an error lint. (Avoiding this was the motivation
for this change.)