CFI: Strip auto traits from Virtual receivers
As the instance being called is behind a vtable, it cannot depend on
auto traits on the receiver (unless the principal trait requires them,
in which case the additional constraint is not needed).

Removing this causes the type signature of the `Virtual` instance to
match the type signature of the `CfiShim`-wrapped entry in the vtable.